+++ packages/accounts-controller/src/AccountsController.test.ts
@@ -1407,10 +1407,7 @@ describe('AccountsController', () => {
           [],
         );
 
-        // First call is 'AccountsController:stateChange'
-        expect(messengerSpy).toHaveBeenNthCalledWith(
-          // 1. AccountsController:stateChange
-          2,
+        expect(messengerSpy).toHaveBeenCalledWith(
           'AccountsController:accountAdded',
           MockExpectedInternalAccountBuilder.from(mockAccount2)
             .setExpectedLastSelectedAsAny()
@@ -1728,10 +1725,7 @@ describe('AccountsController', () => {
           [],
         );
 
-        // First call is 'AccountsController:stateChange'
-        expect(messengerSpy).toHaveBeenNthCalledWith(
-          // 1. AccountsController:stateChange
-          2,
+        expect(messengerSpy).toHaveBeenCalledWith(
           'AccountsController:accountRemoved',
           mockAccount3.id,
         );
@@ -3699,16 +3693,9 @@ describe('AccountsController', () => {
         accountsController.state.internalAccounts.selectedAccount,
       ).toStrictEqual(mockNonEvmAccount.id);
 
-      expect(messengerSpy.mock.calls).toHaveLength(2); // state change and then selectedAccountChange
-
       expect(messengerSpy).not.toHaveBeenLastCalledWith(
         'AccountsController:selectedEvmAccountChange',
         mockNonEvmAccount,
-      );
-
-      expect(messengerSpy).toHaveBeenLastCalledWith(
-        'AccountsController:selectedAccountChange',
-        setExpectedLastSelectedAsAny(mockNonEvmAccount),
       );
     });
   });